move structurehandler and logger into folder and upadate references
